requirements.txt
streamlit 
python-dotenv 
mysql-connector-python 
google-generativeai